backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymysql
python-dotenv
beautifulsoup4
requests
langchain-core
langchain-community
langchain-google-genai
pydantic
